requirements.txt
Flask==0.10.1
Flask-Migrate==1.5.1
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-WTF==0.12
Jinja2==2.8
Mako==1.0.2
MarkupSafe==0.23
SQLAlchemy==1.0.8
Unidecode==0.04.18
WTForms==2.0.2
Werkzeug==0.10.4
alembic==0.8.2
gunicorn==19.3.0
itsdangerous==0.24
psycopg2==2.6.1
python-editor==0.4
python-slugify==1.1.3
wsgiref==0.1.2
